GPU: detect supported platform graphic API version
