+++ src/util/windows/unicode.hpp
@@ -11,6 +11,11 @@ extern "C" {
 
 #include <string_view>
 
+static inline void ffStrbufInitWSV(FFstrbuf* result, const std::wstring_view source)
+{
+    return ffStrbufInitNWS(result, (uint32_t) source.size(), source.data());
+}
+
 static inline void ffStrbufSetWSV(FFstrbuf* result, const std::wstring_view source)
 {
     return ffStrbufSetNWS(result, (uint32_t) source.size(), source.data());
